CSS page
This is some of the CSS I played around with for the page. Feel free to copy it, play around with it, change it and post it as you wish. I hope the link works, haha, otherwise I'll have to try to figure out how to push it to your GitHub via Command Line :sos: 
https://vscode.dev/github/lindsay-evans/odin-recipes/blob/863cd2003f4cf6fae2b145ee7c8bbcbd24467e3aProjects/style.css#L43